switch to useNavigate hook
