+++ client/src/pages/Login/Login.story.tsx
@@ -1,6 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
 import Login from "./Login"
+import { MemoryRouter } from "react-router-dom"
 
 const meta: Meta<typeof Login> = {
   component: Login
@@ -9,4 +10,12 @@ const meta: Meta<typeof Login> = {
 export default meta
 type Story = StoryObj<typeof Login>
 
-export const DefaultLoginPage: Story = {}
+export const DefaultLoginPage: Story = {
+  decorators: [
+    (Story) => (
+      <MemoryRouter>
+        <Story />
+      </MemoryRouter>
+    )
+  ]
+}
